pyproject.toml
[project]
name = "ff-winners-and-losers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "accelerate>=1.10.1",
    "diffusers==0.32.2",
    "espn-api>=0.45.1",
    "google-cloud-aiplatform>=1.121.0",
    "huggingface-hub>=0.35.3",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "protobuf>=6.32.1",
    "pytest>=8.4.2",
    "safetensors>=0.6.2",
    "scipy>=1.13.1",
    "sentencepiece>=0.2.1",
    "transformers==4.49",
    "urllib3<2",
    "vertexai>=1.43.0",
]
